Backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pillow==10.1.0
numpy==1.24.3
werkzeug==3.0.1
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0
scipy==1.11.3
opencv-python==4.8.1.78
accelerate==0.24.0